pyproject.toml
[project]
name = "vibecoder"
version = "0.0.1"
description = "A tool to create and manage GitHub Codespaces with Claude Code integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
]

[project.scripts]
vibecoder = "vibecoder.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vibecoder"]

[tool.uv]
dev-dependencies = [
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

